move applicant navbar into applicantLayout.java and shared in all pages
